Allow JsonNode properties to be treated as values
We need JSON action arguments and response attributes in
some cases: especially for being able to port older services
to Apache Isis while keeping the structure of the JSON which
is posted by hand-written clients.

These changes make it possible to decode and encode JSON
arguments as action parameters and parts of response view
models.